+++ src/components/atoms/Icon.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import { createUseStyles, useTheme } from 'react-jss'
 import clsx from 'clsx'
 
 interface IconProps {
@@ -10,9 +11,19 @@ const defaultProps: IconProps = {
   name: 'bug'
 }
 
+const useStyles = createUseStyles((theme) => ({
+  Icon: {
+    fontSize: '12px',
+  },
+}))
+
 export const Icon: React.FC<IconProps> = ({ name, className }) => {
+  const theme = useTheme()
+  const classes = useStyles({ theme })
+
   const iconStyles = clsx(
-    'fa',
+    classes.Icon,
+    'fas',
     `fa-${name}`,
     className && `${className}`
   )